+++ examples/comments/create_comment.go
@@ -32,7 +32,7 @@ func main() {
 	postID := "a7b2512f-7ee7-4d7f-8224-8d25e94bf0ed"
 
 	// Создаём простой текстовый комментарий
-	filePath := "/home/user/Downloads/music/fakemink_night_blooming_Jasmine.mp3"
+	filePath := "/home/user/Pictures/cat.gif"
 	comment, err := client.Comments.CreateComment(ctx, postID, "Fakemink Night, Blooming Jasmine", filePath)
 	if err != nil {
 		log.Fatal(err)